[refactor] better test cases management
