+++ src/test/java/scs/comp5903/cucumber/integration/cms/util/SeleniumFactory.java
@@ -1,4 +1,4 @@
-package scs.comp5903.cucumber.integration.cmsutil;
+package scs.comp5903.cucumber.integration.cms.util;
 
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.openqa.selenium.WebDriver;
