+++ src/test/java/scs/comp5903/cucumber/util/ReflectionUtilTest.java
@@ -1,7 +1,7 @@
 package scs.comp5903.cucumber.util;
 
 import org.junit.jupiter.api.Test;
-import scs.comp5903.cucumber.integration.cmsutil.SeleniumFactory;
+import scs.comp5903.cucumber.integration.cms.util.SeleniumFactory;
 import scs.comp5903.cucumber.integration.samplestepdef.RummikubDummySeparatedJStepDefs;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
